Issue 44605: Metrics for pipeline jobs
