+++ api/src/org/labkey/api/util/MaintenancePipelineJob.java
@@ -43,8 +43,8 @@ class MaintenancePipelineJob extends PipelineJob
 
     MaintenancePipelineJob(ViewBackgroundInfo info, PipeRoot pipeRoot, Collection<MaintenanceTask> tasks)
     {
-        super(null, info, pipeRoot);
-        setLogFile(new File(pipeRoot.getLogDirectory(), FileUtil.makeFileNameWithTimestamp("system_maintenance", "log")));
+        super("SystemMaintenance", info, pipeRoot);
+        setLogFile(new File(pipeRoot.getLogDirectory(), FileUtil.makeFileNameWithTimestamp("system_maintenance", "log")).toPath());
         _tasks = tasks;
     }
