+++ pipeline/src/org/labkey/pipeline/PipelineModule.java
@@ -25,6 +25,9 @@ import org.labkey.api.collections.CaseInsensitiveHashSet;
 import org.labkey.api.data.Container;
 import org.labkey.api.data.ContainerManager;
 import org.labkey.api.data.RuntimeSQLException;
+import org.labkey.api.data.SQLFragment;
+import org.labkey.api.data.SqlExecutor;
+import org.labkey.api.data.SqlSelector;
 import org.labkey.api.files.FileContentService;
 import org.labkey.api.files.TableUpdaterFileListener;
 import org.labkey.api.module.ModuleContext;
@@ -42,6 +45,8 @@ import org.labkey.api.pipeline.file.PathMapperImpl;
 import org.labkey.api.pipeline.trigger.PipelineTriggerRegistry;
 import org.labkey.api.pipeline.trigger.PipelineTriggerType;
 import org.labkey.api.security.User;
+import org.labkey.api.usageMetrics.UsageMetricsProvider;
+import org.labkey.api.usageMetrics.UsageMetricsService;
 import org.labkey.api.util.ContextListener;
 import org.labkey.api.util.PageFlowUtil;
 import org.labkey.api.util.emailTemplate.EmailTemplateService;
@@ -88,7 +93,9 @@ import java.beans.PropertyChangeEvent;
 import java.io.IOException;
 import java.util.Collection;
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Set;
 
 /**
@@ -203,6 +210,23 @@ public class PipelineModule extends SpringModule implements ContainerManager.Con
         }
 
         AuditLogService.get().registerAuditType(new ProtocolManagementAuditProvider());
+
+        UsageMetricsService.get().registerUsageMetrics("Pipeline", new UsageMetricsProvider()
+        {
+            @Override
+            public Map<String, Object> getUsageMetrics()
+            {
+                Map<String, Long> jobCounts = new HashMap<>();
+                SQLFragment sql = new SQLFragment("SELECT COUNT(*) AS JobCount, COALESCE(Provider, 'Unknown') AS Provider FROM ");
+                sql.append(PipelineSchema.getInstance().getTableInfoStatusFiles(), "sf");
+                sql.append(" GROUP BY Provider");
+
+                new SqlSelector(PipelineSchema.getInstance().getSchema(), sql).forEach(rs ->
+                        jobCounts.put(rs.getString("Provider"), rs.getLong("JobCount")));
+
+                return Collections.singletonMap("jobCounts", jobCounts);
+            }
+        });
     }
 
     @Override
