+++ pipeline/src/org/labkey/pipeline/analysis/CommandTaskImpl.java
@@ -511,78 +511,6 @@ public class CommandTaskImpl extends WorkDirectoryTask<CommandTaskImpl.Factory> 
         }
     }
 
-    // CONSIDER: Move to ScriptTaskImpl -- it's the only usage of this method
-    protected void writeTaskInfo(File file, RecordedAction action) throws IOException
-    {
-        List<String> columns = Arrays.asList("Name", "Value");
-        RowMapFactory<Object> factory = new RowMapFactory<>(columns);
-        List<Map<String, Object>> rows = new ArrayList<>();
-
-        // Job information
-        rows.add(factory.getRowMap("provider", getJob().getProvider()));
-        rows.add(factory.getRowMap("description", getJob().getDescription()));
-        //rows.add(factory.getRowMap("taskPipelineId", getJob().getTaskPipelineId()));
-        rows.add(factory.getRowMap("jobGUID", getJob().getJobGUID()));
-        rows.add(factory.getRowMap("parentGUID", getJob().getParentGUID()));
-        rows.add(factory.getRowMap("splitJob", getJob().isSplitJob()));
-
-        rows.add(factory.getRowMap("baseUrl", AppProps.getInstance().getBaseServerUrl()));
-        rows.add(factory.getRowMap("contextPath", AppProps.getInstance().getContextPath()));
-        rows.add(factory.getRowMap("containerPath", getJob().getContainer().getPath()));
-        rows.add(factory.getRowMap("containerId", getJob().getContainer().getEntityId()));
-        rows.add(factory.getRowMap("user", getJob().getUser().getEmail()));
-
-        PipeRoot pipeRoot = getJob().getPipeRoot();
-        rows.add(factory.getRowMap("pipeRoot", getJob().getPipeRoot().getRootPath()));
-
-        // FileAnalysisJobSupport properties
-        FileAnalysisJobSupport support = getJobSupport();
-        rows.add(factory.getRowMap("protocol", support.getProtocolName()));
-        rows.add(factory.getRowMap("baseName", support.getBaseName()));
-        rows.add(factory.getRowMap("joinedBaseName", support.getJoinedBaseName()));
-        rows.add(factory.getRowMap("analysisDirectory", rewritePath(support.getAnalysisDirectory().toString())));
-        rows.add(factory.getRowMap("dataDirectory", rewritePath(support.getDataDirectory().toString())));
-
-        // Task information
-        rows.add(factory.getRowMap("taskId", _factory.getId()));
-
-        // Write out the known inputs to this task
-        for (RecordedAction.DataFile inputFile : action.getInputs())
-        {
-            String role = inputFile.getRole();
-            if (role == null)
-                continue;
-
-            URI uri = inputFile.getURI();
-            File f = new File(uri);
-            if (f.exists())
-            {
-                String inputPath = _wd.getRelativePath(f);
-                rows.add(factory.getRowMap(role, inputPath));
-            }
-        }
-
-        for (Map.Entry<String, TaskPath> entry : _factory.getOutputPaths().entrySet())
-        {
-            String key = entry.getKey();
-            TaskPath path = entry.getValue();
-
-            // CONSIDER: Include the TaskPath information (optional, etc.)
-
-            String[] outputPaths = getProcessPaths(WorkDirectory.Function.output, key);
-            for (String outputPath : outputPaths)
-            {
-                rows.add(factory.getRowMap(key, outputPath, path.getType().getDefaultSuffix()));
-            }
-        }
-
-        try (TSVMapWriter tsvWriter = new TSVMapWriter(columns, rows))
-        {
-            tsvWriter.setHeaderRowVisible(false);
-            tsvWriter.write(file);
-        }
-    }
-
     /**
      * The parameter replacement map created as the task is run and used
      * to replace tokens the script or the command line before executing it.
