+++ study/src/org/labkey/study/importer/StudyImportJob.java
@@ -76,7 +76,7 @@ public class StudyImportJob extends PipelineJob implements StudyJobSupport, Stud
     // Handles all four study import tasks: initial task, dataset import, specimen import, and final task
     public StudyImportJob(Container c, User user, ActionURL url, Path studyXml, String originalFilename, BindException errors, PipeRoot pipeRoot, ImportOptions options)
     {
-        super(null, new ViewBackgroundInfo(c, user, url), pipeRoot);
+        super("StudyImport", new ViewBackgroundInfo(c, user, url), pipeRoot);
         _originalFilename = originalFilename;
         String baseLogFileName = StudyPipeline.getLogFilename(studyXml.getParent().resolve("study_load"));
         setupLocalDirectoryAndJobLog(pipeRoot, StudyModule.MODULE_NAME, baseLogFileName);
